Separated passkey out. Added wallet functions for pk functions called from UI Closes #298
